replicate Zoombox style from WPF Toolkit, except the Drag/Zoom Modifier cursors
I don't want to require hotkeys for panning/zooming, but with default settings
it simply keeps showing either default cursors (since AreXxxActive properties
are true every time in this case).

To make this work, override the style but include the other setters to keep
the rest of the functionality (such as the rectangle zoom or the panning
cursor while actually panning).